Add defaults for transitions/dependency relations
